# docs/swagger.yaml
basePath: /api
definitions:
...
  models.NullInt64:
    properties:
      int64:
        format: int64
        type: integer
      valid:
        description: Valid is true if Int64 is not NULL
        type: boolean
    type: object
  models.NullString:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
  models.NullTime:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Order:
    properties:
      client_id:
        type: integer
      client_name:
        type: string
      created_at:
        type: string
      due_date:
        $ref: '#/definitions/models.NullTime'
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      proposal_id:
        $ref: '#/definitions/models.NullInt64'
      proposal_reference:
        $ref: '#/definitions/models.NullString'
      reference:
        type: string
      section_id:
        $ref: '#/definitions/models.NullInt64'
      section_name:
        $ref: '#/definitions/models.NullString'
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.OrderCreate:
    properties:
      client_id:
        type: integer
      due_date:
        $ref: '#/definitions/models.NullTime'
      proposal_id:
        $ref: '#/definitions/models.NullInt64'
      reference:
        type: string
      section_id:
        $ref: '#/definitions/models.NullInt64'
      status:
        type: string
    required:
    - client_id
    - reference
    type: object
  models.OrderItem:
    properties:
      id:
        type: integer
      notes:
        type: string
      order_id:
        type: integer
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: number
      unit:
        type: string
      unit_price:
        type: number
    type: object
  models.Product:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      unit:
        type: string
      unit_price:
        type: number
    type: object
  models.ProductCreate:
    properties:
      description:
        type: string
      name:
        type: string
      unit:
        type: string
      unit_price:
        type: number
    required:
    - name
    - unit
    - unit_price
    type: object
...
  models.ProductSearchResult:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:5000
info:
  contact:
    email: support@aura-erp.com
    name: API Support
    url: http://www.aura-erp.com/support
  description: This is a powerful ERP backend for Aura.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Aura ERP API
  version: "1.0"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: Get all orders in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order
      parameters:
      - description: Order details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.OrderCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an order
      tags:
      - orders
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an order
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an order
      tags:
      - orders
    get:
      consumes:
      - application/json
      description: Get order by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get order by ID
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Update an order
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an order
      tags:
      - orders
  /orders/items:
    get:
      consumes:
      - application/json
      description: Get all order items in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OrderItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List order items
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order item
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrderItem'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an order item
      tags:
      - orders
  /orders/items/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an order item
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrderItem'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an order item
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Update an order item
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrderItem'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an order item
      tags:
      - orders
  /products:
    get:
      consumes:
      - application/json
      description: Get all products in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.ProductCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a product
      tags:
      - products
  /products/search:
    get:
      consumes:
      - application/json
      description: Search products by name
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: 25
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProductSearchResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search products
      tags:
      - products
...
swagger: "2.0"